Commit to giving spaces owners, rather than marks
I think I see what I did. I was thinking of a space as marked or unmarked. But I
wanted a space to be able to participate in tactical information, so it should
know its owner. The owner, and the mark, are overlapping but non-equal pieces of
information.

Now that I've committed to one or the other, the two test failures are BOTH
related to problems with using Symbols to represent players, at least that's my
interpretation of this:

Failures:

  1) Space having been marked should know who occupies it (player)
     Failure/Error: expect(@space.owner).to eq(:p1)

     NoMethodError:
       undefined method `owner' for #<Space:0x000000020412d8 @owner=:p1>
     # ./spec/space_spec.rb:40:in `block (3 levels) in <top (required)>'

  2) Space having been marked should know what occupies it (glyph)
     Failure/Error: @owner.glyph

     NoMethodError:
       undefined method `glyph' for :p1:Symbol
     # ./lib/space.rb:28:in `glyph'
     # ./spec/space_spec.rb:51:in `block (3 levels) in <top (required)>'