Delete debug tests; improve UX by hardcoding best first move
A "subgame perfect" strategy is one that reaches a Nash equilibrium in all
subgames, not just those you reach by playing a rational opponent. The computer
player isn't subgame perfect; but it's a little closer to being so if it chooses
the center square as a first move, and that means that the console player has
much, much less time to wait to play their first move.